scripts/pyproject.toml
[project]
name = "jakes-dad-scripts"
version = "1.0.0"
description = "Utility scripts for managing Jakes Dad league data"
requires-python = ">=3.10"
dependencies = [
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
sync-data = "jakes_dad_scripts.sync_league_data:main"
update-stats = "jakes_dad_scripts.update_player_stats:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
